feat(EMS-508): split up single contract policy validation e2e tests
